refactor: split CreateShopMaintenanceCommands into 3 focused classes
CreateShopMaintenanceCommands: 1844 -> 320 lines (router only). CreateShopResetCommands (810 lines): prepareUninstall, resetLiveState, graph-aware cancel helpers. CreateShopDiagnosticCommands (242 lines): runLiveTest, selectLiveTestStack, findLiveTestTargetRequester. CreateShopTestHarnessCommands (573 lines): auto_test_harness commands, lost-package harness operations. No logic changes.
